Got everything finally working

+++ run.sh
@@ -1,11 +1,40 @@
+#!/bin/bash
 
-rm -rf build
-mkdir build
+clean=false
 
-conan install . --output-folder=build --build=missing
+while getopts "drvc" opt; do
+  case $opt in
+    d)
+      BUILD_TYPE=Debug
+      ;;
+    r)
+      BUILD_TYPE=Release
+      ;;
+    v)
+      verbose=true
+      ;;
+    c)
+      clean=true
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG"
+      ;;
+  esac
+done
+
+# if ["$clean" == true]; then
+if $clean; then
+  rm -rf build
+  mkdir build
+  conan install . \
+    --output-folder=build \
+    --build=missing \
+    --settings=build_type=$BUILD_TYPE \
+    --profile=default
+fi
 
 cd build
-cmake .. -DCMAKE_TOOLCHAIN_FILE=conan_toolchain.cmake -DCMAKE_BUILD_TYPE=Release
+cmake .. -DCMAKE_TOOLCHAIN_FILE=conan_toolchain.cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE
 
 cmake --build . --verbose
-ctest -C Release --verbose
+ctest -C $BUILD_TYPE --verbose
